common: adjust sway_log for 32-bit architectures after d7d29fe546ab
common/ipc-client.c:78:4: error: format specifies type 'long' but the argument has type 'time_t' (aka 'int') [-Werror,-Wformat]
                        tv.tv_sec, tv.tv_usec);
                        ^~~~~~~~~
include/log.h:40:74: note: expanded from macro 'sway_log'
        _sway_log(verb, "[%s:%d] " fmt, _sway_strip_path(__FILE__), __LINE__, ##__VA_ARGS__)
                                                                                ^~~~~~~~~~~

+++ common/ipc-client.c
@@ -75,7 +75,7 @@ bool ipc_set_recv_timeout(int socketfd, struct timeval tv) {
 		return false;
 	}
 	sway_log(SWAY_DEBUG, "ipc recv timeout set to %ld.%06ld",
-			tv.tv_sec, tv.tv_usec);
+			(long)tv.tv_sec, (long)tv.tv_usec);
 	return true;
 }
